🧹 T83XJA quality: isolate evaluator failure recovery
