+++ .agentplane/tasks/202609121424-T83XJA/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202609121424-T83XJA/implement-durable-0-7-9-usage-cost-and-latency-a",
   "created_at": "2026-09-12T20:52:29.086Z",
-  "diffstat_sha256": "sha256:ed793cd825762e666fca3e9a50389865f1f747057e87d73c48e9399d0d28dc41",
+  "diffstat_sha256": "sha256:a0c4c39809ab148b190a3f4ee9de6c43d6efddfdef6bf3c4ebe9fd90f4a5ecbc",
   "last_verified_at": "2026-09-13T01:20:05.778Z",
   "last_verified_diffstat_sha256": "sha256:18b0cc41cbe6605ffbbebdf36a8ba44f1097b4398490771f69429af39490c68c",
   "pr_number": 5944,
